Struct redfa::derivatives::Derivatives

pub struct Derivatives<T, R> {
    pub d: Vec<(Vec<T>, R)>,
    pub rest: R,
}

The set of some object's derivatives with respect to an alphabet T.

Fields
d: Vec<(Vec<T>, R)>

Holds a set of pairs (chars, derivative), meaning that the derivative with respect to any element of chars is derivative.

rest: R

The derivative with respect to any character not listed in d.

Methods

impl<T, R> Derivatives<T, R>

fn map<F: FnMut(R) -> R>(self, f: F) -> Derivatives<T, R>