Enum redfa::regex::Regex

pub enum Regex<T> {
    Null,
    Empty,
    Except(Vec<T>),
    Alt(Vec<T>, Vec<Regex<T>>),
    And(Vec<Regex<T>>),
    Not(Box<Regex<T>>),
    Cat(Vec<Regex<T>>),
    Kleene(Box<Regex<T>>),
}

A regular expression over the alphabet T.

Variants
Null

The null set. This never matches anything.

Empty

The empty string (matches exactly "").

Except(Vec<T>)

Matches any single character except the listed ones.

Alt(Vec<T>, Vec<Regex<T>>)

Alternation (also known as disjunction). Matches any of the contained characters, as well as any string matched by a contained regex.

And(Vec<Regex<T>>)

Conjunction. Matches iff all contained regexes match.

Not(Box<Regex<T>>)

Negation. Matches iff the contained regex does not match.

Cat(Vec<Regex<T>>)

Concatenation. Matches iff the contained regexes match in sequence.

Kleene(Box<Regex<T>>)

Kleene closure. Matches zero or more repetitions of the contained regex.

Methods

impl<T> Regex<T>

fn nullable(&self) -> bool

Tests whether a regular expression is nullable, i.e. whether it matches the empty string.